5100: moved changes from old trigger branch
