+++ src/ducks/Signals/signalFormManager/signalCrudForm/signal/TriggerForm.js
@@ -13,8 +13,6 @@ import {
   fetchHistorySignalPoints,
   removeTrigger
 } from '../../../common/actions'
-import FormikInput from '../../../../../components/formik-santiment-ui/FormikInput'
-import FormikSelect from '../../../../../components/formik-santiment-ui/FormikSelect'
 import FormikCheckboxes from '../../../../../components/formik-santiment-ui/FormikCheckboxes'
 import FormikToggle from '../../../../../components/formik-santiment-ui/FormikToggle'
 import FormikEffect from '../../../../../components/formik-santiment-ui/FormikEffect'
@@ -22,88 +20,16 @@ import { TriggerFormHeader } from '../header/TriggerFormHeader'
 import styles from './TriggerForm.module.scss'
 
 import {
-  DAILY_ACTIVE_ADDRESSES,
   PRICE_PERCENT_CHANGE,
-  PRICE_VOLUME_DIFFERENCE,
-  METRICS,
-  PRICE_TYPES,
-  METRICS_DEPENDENCIES,
   METRIC_DEFAULT_VALUES,
-  FREQUENCY_TYPES,
-  DEFAULT_FORM_META_SETTINGS,
-  ASSETS_FILTERS,
-  frequencyTymeValueBuilder,
-  getTypeByMetric,
-  getFrequencyTimeType,
-  getFrequencyTimeValues,
-  getNearestFrequencyTimeValue,
-  getNearestFrequencyTypeValue,
-  PRICE_ABSOLUTE_CHANGE_SINGLE_BORDER,
-  PRICE_ABSOLUTE_CHANGE_DOUBLE_BORDER
-} from '../../../utils/utils'
+  DEFAULT_FORM_META_SETTINGS
+} from '../../../utils/constants'
 
-const REQUIRED_MESSAGE = 'Required'
-const MUST_BE_MORE_ZERO_MESSAGE = 'Must be more 0'
-
-const validate = values => {
-  let errors = {}
-
-  if (
-    values.type.value === DAILY_ACTIVE_ADDRESSES ||
-    values.type.value === PRICE_PERCENT_CHANGE
-  ) {
-    if (!values.percentThreshold) {
-      errors.percentThreshold = REQUIRED_MESSAGE
-    } else if (values.percentThreshold <= 0) {
-      errors.percentThreshold = MUST_BE_MORE_ZERO_MESSAGE
-    }
-    if (!values.timeWindow) {
-      errors.timeWindow = REQUIRED_MESSAGE
-    } else if (values.timeWindow <= 0) {
-      errors.timeWindow = MUST_BE_MORE_ZERO_MESSAGE
-    }
-  }
-
-  if (values.type.value === PRICE_ABSOLUTE_CHANGE_SINGLE_BORDER) {
-    if (!values.absoluteThreshold) {
-      errors.absoluteThreshold = REQUIRED_MESSAGE
-    }
-  }
-
-  if (values.type.value === PRICE_ABSOLUTE_CHANGE_DOUBLE_BORDER) {
-    if (!values.absoluteBorderLeft) {
-      errors.absoluteBorderLeft = REQUIRED_MESSAGE
-    }
-    if (!values.absoluteBorderRight) {
-      errors.absoluteBorderRight = REQUIRED_MESSAGE
-    }
-  }
-
-  if (values.type.value === PRICE_VOLUME_DIFFERENCE) {
-    if (!values.threshold) {
-      errors.threshold = REQUIRED_MESSAGE
-    } else if (values.threshold < 0) {
-      errors.threshold = MUST_BE_MORE_ZERO_MESSAGE
-    }
-  }
-  if (values.channels && values.channels.length === 0) {
-    errors.channels = 'You must setup notification channel'
-  }
-
-  if (!values.frequencyType || !values.frequencyType.value) {
-    errors.frequencyType = REQUIRED_MESSAGE
-  }
-
-  if (!values.frequencyTimeValue || !values.frequencyTimeValue.value) {
-    errors.frequencyTimeValue = REQUIRED_MESSAGE
-  }
-
-  if (!values.frequencyTimeType || !values.frequencyTimeType.value) {
-    errors.frequencyTimeType = REQUIRED_MESSAGE
-  }
-
-  return errors
-}
+import { validateTriggerForm } from '../../../utils/utils'
+import { TriggerFormAssetWallet } from '../formParts/TriggerFormAssetWallet'
+import { TriggerFormMetricValues } from '../formParts/TriggerFormMetricValues'
+import { TriggerFormMetricTypes } from '../formParts/TriggerFormMetricTypes'
+import { TriggerFormFrequency } from '../formParts/TriggerFormFrequency'
 
 const propTypes = {
   onSettingsChange: PropTypes.func.isRequired,
@@ -125,8 +51,13 @@ export const TriggerForm = ({
   removeSignal,
   onRemovedSignal
 }) => {
+  const formMetric =
+    metaFormSettings && metaFormSettings.metric
+      ? metaFormSettings.metric.value.value
+      : PRICE_PERCENT_CHANGE
+
   metaFormSettings = { ...DEFAULT_FORM_META_SETTINGS, ...metaFormSettings }
-  settings = { ...METRIC_DEFAULT_VALUES[PRICE_PERCENT_CHANGE], ...settings }
+  settings = { ...METRIC_DEFAULT_VALUES[formMetric], ...settings }
 
   settings = {
     target: metaFormSettings.target.value
@@ -141,17 +72,12 @@ export const TriggerForm = ({
     signalType: metaFormSettings.signalType.value
       ? metaFormSettings.signalType.value
       : settings.signalType,
+    address: metaFormSettings.address,
     ...settings
   }
 
   const [initialValues, setInitialValues] = useState(settings)
   const [showTrigger, setShowTrigger] = useState(true)
-
-  const defaultSignalType = metaFormSettings.signalType
-  const defaultAsset = metaFormSettings.target
-  const defaultMetric = metaFormSettings.metric
-  const defaultType = metaFormSettings.type
-  const defaultFrequencyType = metaFormSettings.frequencyType
 
   const setDefaultPriceValues = values => {
     const newValues = { ...values, ...METRIC_DEFAULT_VALUES[values.type.value] }
@@ -172,12 +98,13 @@ export const TriggerForm = ({
       initialValues={initialValues}
       isInitialValid
       enableReinitialize
-      validate={validate}
+      validate={validateTriggerForm}
       onSubmit={values => {
         onSettingsChange(values)
       }}
     >
       {({
+        values,
         values: {
           metric,
           type,
@@ -198,7 +125,10 @@ export const TriggerForm = ({
         <Form className={styles.TriggerForm}>
           <FormikEffect
             onChange={(current, prev) => {
-              if (current.values.type.value !== prev.values.type.value) {
+              if (
+                !prev.values.type ||
+                current.values.type.value !== prev.values.type.value
+              ) {
                 setDefaultPriceValues(current.values)
                 validateForm()
               }
@@ -216,212 +146,32 @@ export const TriggerForm = ({
 
             {showTrigger && (
               <div className={styles.Trigger}>
-                <div className={styles.row}>
-                  <div className={styles.Field}>
-                    <label>Type</label>
-                    <FormikSelect
-                      name='signalType'
-                      isDisabled={defaultSignalType.isDisabled}
-                      defaultValue={defaultSignalType.value.value}
-                      placeholder='Pick signal type'
-                      options={ASSETS_FILTERS}
-                    />
-                  </div>
-                  <div className={styles.Field}>
-                    <FormikSelect
-                      name='target'
-                      isDisabled={defaultAsset.isDisabled}
-                      defaultValue={defaultAsset.value.value}
-                      placeholder='Pick an asset'
-                      options={allProjects.map(asset => ({
-                        label: asset.slug,
-                        value: asset.slug
-                      }))}
-                    />
-                  </div>
-                </div>
+                <TriggerFormMetricTypes
+                  metaFormSettings={metaFormSettings}
+                  setFieldValue={setFieldValue}
+                  metric={metric}
+                />
 
-                <div className={styles.row}>
-                  <div className={styles.Field}>
-                    <label>Metrics</label>
-                    <div>
-                      <FormikSelect
-                        name='metric'
-                        isClearable={false}
-                        isDisabled={defaultMetric.isDisabled}
-                        defaultValue={defaultMetric.value}
-                        isSearchable
-                        placeholder='Choose a metric'
-                        options={METRICS}
-                        onChange={newMetric => {
-                          metric &&
-                            newMetric.value !== metric.value &&
-                            setFieldValue('type', getTypeByMetric(newMetric))
-                        }}
-                      />
-                    </div>
-                  </div>
-                  {PRICE_TYPES[(metric || {}).value] &&
-                    PRICE_TYPES[(metric || {}).value].length > 1 && (
-                    <div className={styles.Field}>
-                      <FormikSelect
-                        name='type'
-                        isClearable={false}
-                        isSearchable
-                        isDisabled={defaultType.isDisabled}
-                        defaultValue={defaultType.value}
-                        placeholder='Choose a type'
-                        options={PRICE_TYPES[metric.value]}
-                        isOptionDisabled={option => !option.value}
-                      />
-                    </div>
-                  )}
-                </div>
-                <div className={styles.row}>
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes(
-                      'absoluteBorderLeft'
-                    ) && (
-                    <div className={styles.Field}>
-                      <label>Select channel borders</label>
-                      <FormikInput
-                        name='absoluteBorderLeft'
-                        type='number'
-                        max={absoluteBorderRight}
-                        placeholder='Left border'
-                      />
-                    </div>
-                  )}
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes(
-                      'absoluteBorderRight'
-                    ) && (
-                    <div className={styles.Field}>
-                      <FormikInput
-                        name='absoluteBorderRight'
-                        min={absoluteBorderLeft}
-                        type='number'
-                        placeholder='Right border'
-                      />
-                    </div>
-                  )}
+                <TriggerFormAssetWallet
+                  metric={metric}
+                  metaFormSettings={metaFormSettings}
+                  allProjects={allProjects}
+                  setFieldValue={setFieldValue}
+                />
 
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes(
-                      'absoluteThreshold'
-                    ) && (
-                    <div className={styles.Field}>
-                      <label>Absolute change</label>
-                      <FormikInput
-                        name='absoluteThreshold'
-                        type='number'
-                        placeholder='Absolute change'
-                      />
-                    </div>
-                  )}
+                <TriggerFormMetricValues
+                  type={type}
+                  absoluteBorderLeft={absoluteBorderLeft}
+                  absoluteBorderRight={absoluteBorderRight}
+                />
 
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes(
-                      'percentThreshold'
-                    ) && (
-                    <div className={styles.Field}>
-                      <label>Percentage change</label>
-                      <FormikInput
-                        name='percentThreshold'
-                        type='number'
-                        placeholder='Percentage change'
-                      />
-                    </div>
-                  )}
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes('threshold') && (
-                    <div className={styles.Field}>
-                      <label>Threshold</label>
-                      <FormikInput
-                        name='threshold'
-                        step={0.001}
-                        type='number'
-                        placeholder='Threshold'
-                      />
-                    </div>
-                  )}
-                  {type &&
-                    METRICS_DEPENDENCIES[type.value].includes('timeWindow') && (
-                    <div className={styles.Field}>
-                      <label>Time Window</label>
-                      <div className={styles.timeWindow}>
-                        <div className={styles.timeWindowInput}>
-                          <FormikInput
-                            name='timeWindow'
-                            type='number'
-                            min={0}
-                            placeholder='Time window'
-                          />
-                        </div>
-                        <FormikSelect
-                          name='timeWindowUnit'
-                          className={styles.timeWindowUnit}
-                          clearable={false}
-                          placeholder='Unit'
-                          options={[
-                            { value: 'h', label: 'Hours' },
-                            { value: 'd', label: 'Days' }
-                          ]}
-                        />
-                      </div>
-                    </div>
-                  )}
-                </div>
+                <TriggerFormFrequency
+                  metaFormSettings={metaFormSettings}
+                  setFieldValue={setFieldValue}
+                  frequencyType={frequencyType}
+                  frequencyTimeType={frequencyTimeType}
+                />
 
-                <div className={styles.row}>
-                  <div className={styles.Field}>
-                    <label>Frequency of notifications</label>
-                    <FormikSelect
-                      name='frequencyType'
-                      isClearable={false}
-                      isDisabled={defaultFrequencyType.isDisabled}
-                      defaultValue={defaultFrequencyType.value.value}
-                      isSearchable
-                      placeholder='Choose frequency'
-                      options={FREQUENCY_TYPES}
-                      onChange={frequencyType => {
-                        const newFrequencyTimeType = getNearestFrequencyTypeValue(
-                          frequencyType
-                        )
-                        setFieldValue('frequencyTimeType', newFrequencyTimeType)
-                        setFieldValue(
-                          'frequencyTimeValue',
-                          getNearestFrequencyTimeValue(newFrequencyTimeType)
-                        )
-                      }}
-                    />
-                  </div>
-                  <div className={styles.Field}>
-                    <div className={styles.frequency}>
-                      <FormikSelect
-                        name='frequencyTimeValue'
-                        className={styles.frequencyTimeValue}
-                        isClearable={false}
-                        isDisabled={!frequencyType || !frequencyTimeType}
-                        isSearchable
-                        options={getFrequencyTimeValues(frequencyTimeType)}
-                      />
-                      <FormikSelect
-                        className={styles.frequencyTimeType}
-                        name='frequencyTimeType'
-                        isDisabled={!frequencyType}
-                        isClearable={false}
-                        onChange={frequencyTimeType => {
-                          setFieldValue(
-                            'frequencyTimeValue',
-                            frequencyTymeValueBuilder(1)
-                          )
-                        }}
-                        options={getFrequencyTimeType(frequencyType)}
-                      />
-                    </div>
-                  </div>
-                </div>
                 <div className={styles.row}>
                   <div className={styles.Field}>
                     <div className={styles.isRepeating}>
