+++ src/pages/SonarFeed/SignalDetails.module.scss
@@ -27,7 +27,9 @@
 
 .wrapper {
   width: 437px;
+  // height: 160px;
   max-width: 100%;
+  // max-height: 80%;
   background-color: var(--white);
 }
